# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZHIXU API",
    "version": "0.1.0",
    "description": "System health, Workspace, durable Workflow and M5 Change Control contracts. Apply preflight reads the real Workspace file hash but never writes files or creates Git commits."
  },
  "paths": {
...
    "/api/v1/system/status": {
      "get": {
        "operationId": "getSystemStatus",
        "responses": {
          "200": {
            "description": "Current API, database, Graph, and RAG status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatus"
                }
              }
            }
          },
          "405": {
            "description": "The method is not supported.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "DatabaseStatus": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "unavailable"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SystemStatus": {
        "type": "object",
        "required": [
          "status",
          "version",
          "database",
          "graph",
          "semantic_links",
          "rag",
          "collections",
          "knowledge_health",
          "request_id"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "degraded"
            ]
          },
          "version": {
            "type": "string",
            "minLength": 1
          },
          "database": {
            "$ref": "#/components/schemas/DatabaseStatus"
          },
          "graph": {
            "$ref": "#/components/schemas/GraphCapabilityStatus"
          },
          "semantic_links": {
            "$ref": "#/components/schemas/SemanticLinkCapabilityStatus"
          },
          "rag": {
            "$ref": "#/components/schemas/RAGCapabilityStatus"
          },
          "collections": {
            "$ref": "#/components/schemas/OptionalCapabilityStatus"
          },
          "knowledge_health": {
            "$ref": "#/components/schemas/OptionalCapabilityStatus"
          },
          "request_id": {
            "type": "string",
            "minLength": 1
          }
        },
        "additionalProperties": false
      },
      "GraphCapabilityStatus": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "unavailable"
            ]
          },
          "reason": {
            "type": "string",
            "enum": [
              "graph_dependencies_unavailable"
            ]
          }
        },
        "additionalProperties": false
      },
      "SemanticLinkCapabilityStatus": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "unavailable"
            ]
          },
          "reason": {
            "type": "string",
            "enum": [
              "semantic_link_dependencies_unavailable"
            ]
          }
        },
        "additionalProperties": false
      },
      "RAGCapabilityStatus": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "disabled",
              "unavailable"
            ]
          },
          "reason": {
            "type": "string",
            "enum": [
              "rag_dependencies_unavailable"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "required": [
          "error_code",
          "message",
          "retryable"
        ],
        "properties": {
          "error_code": {
            "type": "string",
            "minLength": 1
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "retryable": {
            "type": "boolean"
          },
          "workflow_run_id": {
            "type": "string",
            "format": "uuid"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": false
      },
      "OptionalCapabilityStatus": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "unavailable"
            ]
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}